% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puntocorte.R
\name{puntocorte}
\alias{puntocorte}
\title{Asistente de punto de corte}
\usage{
puntocorte(...)
}
\value{
Una app de shiny es iniciada
}
\description{
Asistente de punto de corte
}
\examples{
\dontrun{
puntocorte()
}
}
